Simplify Sexps and Datum (#28)
This removes the the Sexpression type and unifies it with the Datum type. The distinction didn't bring any advantages and in fact was quite a bit weird at some points. As are result, we have less types to deal with and some implementations even got simpler.

Not that this would have been the main reason, but the benchmarks show that parsing has improved by 17% as a consequence. 🎉