Expand simple let (#30)
Add expansion for let syntax to the expander. To make this work I started to add the macro expansion facility.
It uses the VM to expand the supplied syntax and thus forms the skeleton for user-defined syntax.
However it currently doesn't take care of hygiene (not required so far) and is probably not supporting all cases yet.

It adds an expander for a simple let, which expands into the corresponding lambda expressions.
At a later point I'll think about whether it's preferable to compile let expressions differently, which would required to preserve them and parse them into an expression.