Clean up storage model (#31)
Properly implement the scheme storage model.
This adds a representation of set-able places, that can be mutated with the set! core form.
It implements vector, list (proper and improper) as having their elements represented as places.
In contrast to scheme, where strings are essentially sequences of settable chars, we left strings as immutable.